Added validation in API and added null check in handler

+++ Lapka.Identity.Application/Commands/Handlers/DeleteShelterHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Convey.CQRS.Commands;
 using Lapka.Identity.Application.Dto;
+using Lapka.Identity.Application.Exceptions;
 using Lapka.Identity.Application.Services;
 using Lapka.Identity.Core.Entities;
 
@@ -18,7 +19,9 @@ namespace Lapka.Identity.Application.Commands.Handlers
         }
         public async Task HandleAsync(DeleteShelter command)
         {
-            Shelter shelter = await _shelterRepository.GetByIdAsync(command.Id); 
+            Shelter shelter = await _shelterRepository.GetByIdAsync(command.Id);
+            if (shelter is null) throw new ValueNotFoundException();
+            
             shelter.Delete();
             
             await _shelterRepository.DeleteAsync(shelter);
